Refactor: Add getClient util function
