+++ src/utils/colonyClient.ts
@@ -1,5 +1,3 @@
-import { AnyColonyClient } from '@colony/colony-js';
-
 import { query } from '~amplifyClient';
 import {
   Colony,
@@ -7,35 +5,8 @@ import {
   GetColonyQuery,
   GetColonyQueryVariables,
 } from '~graphql';
-import networkClient from '~networkClient';
 
-const colonyClientsCache: Record<string, AnyColonyClient> = {};
-
-/**
- * Method attempting to return a cached colony client,
- * or get one from the networkClient if there's no cached entry
- */
-export const getCachedColonyClient = async (
-  colonyAddress: string,
-): Promise<AnyColonyClient | undefined> => {
-  if (colonyAddress in colonyClientsCache) {
-    return colonyClientsCache[colonyAddress];
-  }
-
-  try {
-    const colonyClient = await networkClient.getColonyClient(colonyAddress);
-    colonyClientsCache[colonyAddress] = colonyClient;
-    return colonyClient;
-  } catch (error) {
-    console.error(
-      `Unable to fetch colony client for address: ${colonyAddress}`,
-      error,
-    );
-  }
-
-  return;
-};
-
+// @TODO: Consider refactoring this as it doesn't feel like it's in the right place
 export const getColonyFromDB = async (
   colonyAddress: string,
 ): Promise<Colony | undefined> => {
